requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
pyodbc==4.0.39
python-dotenv==1.0.0
bcrypt==4.0.1
Werkzeug==2.3.7